pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "gradio (>=5.31.0,<6.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "chromadb (>=1.0.12,<2.0.0)",
    "gradio-pdf (>=0.0.22,<0.0.23)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "tavily-python (>=0.7.5,<0.8.0)",
    "crewai (>=0.98.0)",
    "selenium (>=4.33.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

